Delete Singularity recipe, move Python to top level
CONTAINERS UNTESTED

- Renamed Python package to 'epoch_containers'
- Updated linting/formatting/typing
- Updated Dockerfile to match new system [UNTESTED]
- Removed Singularity recipe

The aim here is to use just the Docker container when running
both Singularity and Docker. Ideally, we'll also be able to
merge the two scripts. This will help to reduce the overhead
of maintaining both GHCR and SyLabs container repos.

For now, I've just globbed the various README files that were
floating around into one giga-file. This will need a rewrite
before merging.